edit all mandatory function done but not pass the norm

+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: pnimwata <pnimwata@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/20 12:17:21 by pnimwata          #+#    #+#             */
-/*   Updated: 2022/02/20 12:17:21 by pnimwata         ###   ########.fr       */
+/*   Updated: 2022/02/28 14:38:40 by pnimwata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char	*ft_strdup(char const *s)
 	char	*new_s;
 	size_t	i;
 
-	new_s = (char *)ft_calloc ((ft_strlen(s) + 1), (size_t)sizeof (char));
+	new_s = (char *)ft_calloc((ft_strlen(s) + 1), sizeof (char));
 	i = 0;
 	while (i < (ft_strlen(s) + 1))
 	{
